Disband handler, templater and server
